Move and refactor alienspecies model
