+++ Assessments.Frontend.Web/Controllers/AlienSpeciesController.cs
@@ -5,6 +5,7 @@ using Assessments.Frontend.Web.Infrastructure;
 using Assessments.Frontend.Web.Infrastructure.AlienSpecies;
 using Assessments.Frontend.Web.Models;
 using Assessments.Mapping.AlienSpecies;
+using Assessments.Mapping.AlienSpecies.Model;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using X.PagedList;
