+++ Assessments.Frontend.Web/Views/AlienSpecies/2023/AlienSpeciesDetail.cshtml
@@ -1,5 +1,6 @@
 ﻿@model AlienSpeciesDetailViewModel
-@using Assessments.Frontend.Web.Infrastructure.AlienSpecies
+@using Assessments.Frontend.Web.Infrastructure.AlienSpecies;
+@using Assessments.Mapping.AlienSpecies.Model
 
 @{
     AlienSpeciesAssessment2023 assessment = Model.Assessment;
