+++ Assessments.Frontend.Web/Views/AlienSpecies/2023/ListPartials/_Filters.cshtml
@@ -1,9 +1,9 @@
 @model AlienSpeciesListViewModel;
-@using static Assessments.Frontend.Web.Infrastructure.Constants;
 @using static Assessments.Frontend.Web.Infrastructure.Helpers;
 @using static Assessments.Frontend.Web.Infrastructure.AlienSpecies.FilterHelpers;
 @using System.Linq;
-@using Assessments.Frontend.Web.Infrastructure.AlienSpecies
+@using Assessments.Frontend.Web.Infrastructure.AlienSpecies;
+@using Assessments.Mapping.AlienSpecies.Model
 
 @{
     var AllAreas = Areas.AlienSpecies2023Areas;
