+++ Assessments.Mapping/AlienSpecies/Model/AlienSpeciesAssessment2023PreviousAssessment.cs
@@ -1,0 +1,16 @@
+﻿namespace Assessments.Mapping.AlienSpecies.Model
+{
+    public class AlienSpeciesAssessment2023PreviousAssessment
+    {
+        public int RevisionYear { get; set; } = 2018;
+
+        public string AssessmentId { get; set; }
+
+        public int RiskLevel { get; set; }
+
+        /// <summary>
+        /// Evaluation  'tag' for the species in 2018 or 2012
+        /// </summary>
+        public string MainCategory { get; set; }
+    }
+}
