+++ Assessments.Mapping/AlienSpecies/Model/Enums/AlienSpeciesAssessment2023Category.cs
@@ -1,0 +1,29 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Assessments.Mapping.AlienSpecies.Model.Enums
+{
+    public enum AlienSpeciesAssessment2023Category
+    {
+        // ReSharper disable InconsistentNaming
+        [Display(Name = "Ukjent")]
+        Undefined,
+
+        [Display(Name = "Svært høy risiko")]
+        SE,
+
+        [Display(Name = "Høy risiko")]
+        HI,
+
+        [Display(Name = "Potensielt høy risiko")]
+        PH,
+
+        [Display(Name = "Lav risiko")]
+        LO,
+
+        [Display(Name = "Ingen kjent risiko")]
+        NK,
+
+        [Display(Name = "Ikke risikovurdert")]
+        NR
+    }
+}
