+++ Assessments.Mapping/AlienSpecies/Model/Enums/AlienSpeciesAssessment2023EvaluationContext.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Assessments.Mapping.AlienSpecies.Model.Enums
+{
+    public enum AlienSpeciesAssessment2023EvaluationContext
+    {
+        [Display(Name = "Fastlands-Norge med havområder")]
+        N,
+
+        [Display(Name = "Svalbard med havområder")]
+        S
+    }
+}
